requirements.txt
python-telegram-bot
google-generativeai
chromadb
python-dotenv
langdetect
pytest